refactor wasi environment to pure cli environment adapter
