+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["core", "tools", "secure-agent", "runtime"]
+members = ["core", "tools", "adapters", "secure-agent", "runtime"]
 resolver = "2"
