+++ adapters/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "verity-adapters"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+verity-core = { path = "../core" }
