+++ secure-agent/Cargo.toml
@@ -15,6 +15,7 @@ tracing = "0.1"
 http = "1"
 http-body-util = "0.1"
 bytes = "1"
+verity-adapters = { path = "../adapters" }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 getrandom = "0.2"
