--- secure-agent/src/wasi_environment.rs
@@ -1,17 +1,0 @@
-use verity_core::environment::{Environment, LoggingLevel};
-
-pub struct WasiEnvironment;
-
-impl Environment for WasiEnvironment {
-    fn print(&self, s: &str) {
-        println!("{s}");
-    }
-
-    fn logging_level(&self) -> LoggingLevel {
-        LoggingLevel::Standard
-    }
-
-    fn emit_log(&self, message: &str) {
-        eprintln!("{message}");
-    }
-}
